Added a cross-platform socket wrapper with platform-specific implementations, exposing a common API for socket creation, configuration, and I/O across Windows and POSIX systems
Refactored networking utilities to rely on the new wrapper and adopted fixed-width types for addresses, ports, and timeouts, improving 64-bit correctness and portability

Replaced direct Winsock usage in connection and packet management code with wrapper calls and standardized integer widths, removing platform dependencies and aligning bandwidth calculations with uint32_t semantics

+++ Code/wwnet/BWBalance.h
@@ -36,6 +36,7 @@
 
 #pragma once
 #ifndef _BWBALANCE_H
+#include <cstdint>
 
 class cConnection;
 
@@ -63,7 +64,7 @@ class BandwidthBalancerClass
 
 	private:
 
-		unsigned long Allocate_Bandwidth(float average_priority, int bw_adjust, unsigned long total_server_bbo);
+		uint32_t Allocate_Bandwidth(float average_priority, int bw_adjust, uint32_t total_server_bbo);
 		void Allocate_Client_Structs(int num_structs);
 		void Adjust_Connection_Budget(cConnection *connection);
 
@@ -73,10 +74,10 @@ class BandwidthBalancerClass
 		*/
 		struct ClientInfoStruct {
 			float AveragePriority;
-			unsigned long MaxBpsDown;
-			unsigned long AllocatedBBO;
+			uint32_t MaxBpsDown;
+			uint32_t AllocatedBBO;
 			bool IsLoading;
-			unsigned long ID;
+			uint32_t ID;
 			bool IsDone;
 		};
